fix more shared test setup
