--- client/test/app/intake/testHelpers.js
@@ -1,5 +1,0 @@
-export const newProps = {
-  cancelText: 'cancel',
-  skipText: 'skip',
-  submitText: 'submit'
-};
